+++ src/dynamodb-stream/parser.ts
@@ -1,14 +1,14 @@
 import { DynamoDBStreamEvent, DynamoDBRecord } from 'aws-lambda';
 import { DynamoDB, DynamoDBStreams } from 'aws-sdk';
 
-export class DynamoDBStreamParser {
+export class DynamoDBStreamParser<T> {
   constructor(private event: DynamoDBStreamEvent) {}
 
-  getVersions(): { newVersions: any[]; oldVersions: any[]; versions: Version[] } {
+  getVersions(): { newVersions: any[]; oldVersions: any[]; versions: Version<T>[] } {
     let newVersions = [];
     let oldVersions = [];
-    let versions: Version[] = [];
-    this.event.Records.forEach(record => {
+    let versions: Version<T>[] = [];
+    this.event.Records.forEach((record) => {
       const { newVersion, oldVersion, version } = this.parseVersions(record);
       if (newVersion) {
         newVersions.push(newVersion);
@@ -26,11 +26,11 @@ export class DynamoDBStreamParser {
   private parseVersions(record: DynamoDBRecord) {
     const { eventName } = record;
     const { NewImage, OldImage, Keys } = record.dynamodb;
-    const newVersion = NewImage ? DynamoDB.Converter.unmarshall(NewImage) : undefined;
-    const oldVersion = OldImage ? DynamoDB.Converter.unmarshall(OldImage) : undefined;
+    const newVersion = NewImage ? (DynamoDB.Converter.unmarshall(NewImage) as T) : undefined;
+    const oldVersion = OldImage ? (DynamoDB.Converter.unmarshall(OldImage) as T) : undefined;
     const keys = DynamoDB.Converter.unmarshall(Keys);
     const { tableName, tableArn } = this.getTableInformation(record);
-    const version: Version = { newVersion, oldVersion, keys, tableName, tableArn, eventName };
+    const version: Version<T> = { newVersion, oldVersion, keys, tableName, tableArn, eventName };
     return { newVersion, oldVersion, version };
   }
 
@@ -42,11 +42,11 @@ export class DynamoDBStreamParser {
   }
 }
 
-export interface Version {
-  newVersion: any; // unmarshalled javascript object of new image (if exists) or null
-  oldVersion: any; // unmarshalled javascript object of old image (if exists) or null
+export interface Version<T> {
+  newVersion: T; // unmarshalled javascript object of new image (if exists) or null
+  oldVersion: T; // unmarshalled javascript object of old image (if exists) or null
   keys: any; // unmarshalled javascript object of keys (includes key values)
   tableName: string; // name of the table the object came from
   tableArn: string; // arn of the table the object came from
-  eventName: string; // name of the event (INSERT || MODIFY || REMOVE)
+  eventName: 'INSERT' | 'MODIFY' | 'REMOVE'; // name of the event (INSERT || MODIFY || REMOVE)
 }