+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
+## [3.2.0]&nbsp;&nbsp;(2020-05-02)
+
+### Changed
+
+- DynamoDB stream wrapper accepts optional type generic for better TypeScript support
+- DynamoDB stream wrapper returns string enum for DynamoDB event type instead of 'string'
+
 ## [3.1.3]&nbsp;&nbsp;(2020-03-16)
 
 ### Changed
@@ -176,6 +183,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/), 
 - Update older libraries
 - Now publish from Git tags instead of master pushes
 
+[3.2.0]: https://github.com/manwaring/lambda-wrapper/compare/v3.1.3...v3.2.0
 [3.1.3]: https://github.com/manwaring/lambda-wrapper/compare/v3.1.2...v3.1.3
 [3.1.2]: https://github.com/manwaring/lambda-wrapper/compare/v3.1.1...v3.1.2
 [3.1.1]: https://github.com/manwaring/lambda-wrapper/compare/v3.1.0...v3.1.1